Create YouTube Thumbnails.py
This Python script will take an Inkscape SVG, copy it to a temporary file, change two strings in it, and then call Inkscape to export a PNG with these new strings. This is helpful when generating thumbnails that share a background but have different titles.
It can take a number to generate a series of 'Part #' thumbnails. It can also take a plain text list of titles and place them in each PNG.
It will check if the PNG already exists.