Checks for both versions of the base thumbnail
The script generates two versions of a 'base' thumbnail that does not identify the part. One is a full resolution version to use as a background in OBS and the other is really not used for anything aside from checking file sizes. Anyway, this change has it check for each to match sure that if one is missing, it gets made.